+++ src/main/java/roomescape/ReservationController.java
@@ -30,22 +30,8 @@ public class ReservationController {
     }
 
     @PostMapping("/reservations")
-    public ResponseEntity<Reservation> create(@RequestBody Reservation reservation) {
-        if (reservation.getName().isEmpty()) {
-            return handleNotFoundReservationException(new NotFoundReservationException("Name of Reservation is empty"));
-        }
-        if (reservation.getDate().isEmpty()) {
-            return handleNotFoundReservationException(new NotFoundReservationException("Date of Reservation is empty"));
-        }
-        if (reservation.getTime() == null) {
-            return handleNotFoundReservationException(new NotFoundReservationException("Time of Reservation is empty"));
-        }
-
-        Long id = reservationService.createReservation(reservation);
-        reservation.setId(id);
-
-        return ResponseEntity.created(URI.create("/reservations/" + reservation.getId()))
-                .contentType(MediaType.APPLICATION_JSON).body(reservation);
+    public ResponseEntity<Reservation> create(@RequestBody ReservationAddRequest reservationAddRequest) {
+        return reservationService.createReservation(reservationAddRequest);
     }
 
     @DeleteMapping("/reservations/{id}")
@@ -56,11 +42,6 @@ public class ReservationController {
         } else {
             return ResponseEntity.noContent().build();
         }
-    }
-
-    @ExceptionHandler(NotFoundReservationException.class)
-    public ResponseEntity handleNotFoundReservationException(NotFoundReservationException e) {
-        return ResponseEntity.badRequest().build();
     }
 
 }
